refactor: teamName 패키지 이동 및 출력 메소드 변경
