+++ src/main/java/chess/view/OutputView.java
@@ -1,6 +1,7 @@
 package chess.view;
 
 import chess.controller.ChessBoardFormatter;
+import chess.domain.piece.Team;
 
 import java.util.List;
 
@@ -29,20 +30,20 @@ public final class OutputView {
         System.out.println("abcdefgh");
     }
 
-    public static void printScore(final String team, final double score) {
-        System.out.println(team + "의 점수는 " + score + "점 입니다.");
+    public static void printScore(final Team team, final double score) {
+        System.out.println(TeamName.getNameByTeam(team) + "의 점수는 " + score + "점 입니다.");
     }
 
     public static void printDrawWhenRunning() {
         System.out.println("두 플레이어가 동등하게 겨루고 있습니다.");
     }
 
-    public static void printWinnerWhenRunning(final String winner) {
-        System.out.println(winner + "이(가) 우세합니다.");
+    public static void printWinnerWhenRunning(final Team winner) {
+        System.out.println(TeamName.getNameByTeam(winner) + "이(가) 우세합니다.");
     }
 
-    public static void printWinnerAfterRunning(final String winner) {
-        System.out.println(winner + "이(가) 상태팀의 왕을 잡아 승리하였습니다.");
+    public static void printWinnerAfterRunning(final Team winner) {
+        System.out.println(TeamName.getNameByTeam(winner) + "이(가) 상태팀의 왕을 잡아 승리하였습니다.");
     }
 
     public static void printSaved(final String roomName) {
